[Inductor cutlass backend] Experimental extended Cutlass op generator
So far we have relied on the "generator.py" from
third_party/cutlass/python/cutlass_library/ to generate Cutlass op
configurations to be tried in autotuning. These op configs are not
sufficient to ensure good performance and dtype coverage in all cases.

This diff introduces an extended variant of that generator.py,
provided both as a source file and as a diff ( so that it may be applied
to future versions of cutlass_library.generator )

Test Plan:
 * CI
 * python test/inductor/test_max_autotune.py

ghstack-source-id: a16bd5aee4fbb3243515e8aaadd8cc82ea22ab97
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113932