[Inductor cutlass backend] Allow to fall back to non-fuseable ops
So far, when only the Cutlass GEMM backend was enabled, and the
config.cuda.cutlass_only_evt_capable_ops option was enabled, it could happen
that for some input combinations, there were no Cutlass ops that could handle
it, since, for example, there are no EVT-capable fp32 GEMM ops when both
operands A and B are row-major.

This diff changes the behavior such that in the case that
the mentioned config option is renamed to config.cuda.cutlass_prefer_evt_capable_ops, and now allows to fall back to non-evt
capable ops in the case that not even a single EVT-capable one could be found..

In the case that no GEMM operand of the selected backend can be found at all, the
ATen backend is used as fallback.

ghstack-source-id: 1cd4b645335f71cf8fa5063a73ac3fe7e292f562
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114075