[AOTInductor] Double buffering for Weights
Summary:
This adds function to model container doing weight swapping with double
buffering.

There are 2 parts for double buffering
a) Write constants into inactive buffer
b) Swap active buffer

For (a), we write the constants into the buffer that's currently not in
use, and store the information in both constants map and the
corresponding constant array to read.
For (b), we obtain the lock, and activate the constant map/constant
array that is inactive, and flag the one that's currently in use to
inactive.

Test Plan:
test/cpp/aot_inductor/test.cpp

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 364958194906c10a804015d2e057b65990846d27
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114446