[Cutlass inductor backend] Cutlass GEMM size threshold
Cutlass backend GEMMs are comparatively expensive to compile. So they should only be applied to sufficiently large GEMMs. This small diff introduces a new torch._inductor.config option called "cuda.cutlass_backend_min_gemm_size" which introduces a threshold for the size of GEMM problems that the Cutlass backend will be considered for.

ghstack-source-id: 046a208b885ad6ae9b5fbbef6cdbb2274ff0fcc1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113569