[Inductor cutlass backend] Retuning after fusion
After epilogue fusion, the previously fastest kernels might not be fastest anymore.
More importantly even, Cutlass epilogue fusion can lead to kernels that fail to launch
due to lack of shared memory. A solution to both is to re-tune the kernel in question
when a fusion happens, and possibly even before a fusion decision is being made.

ghstack-source-id: 08e1ff024f39d5708758ecb49f43ff8301473e4f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115173