[funcol] Directly import DeviceMesh to avoid circular dependency
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115649

This diff aims to directly import DeviceMesh from torch.distributed.device_mesh instead of importing it from dist._tensor. This is done to avoid a circular dependency issue. The code changes in each file of the diff are as follows:

- torch/distributed/_functional_collectives.py: import DeviceMesh from torch.distributed instead of dist._tensor.


Overall, this diff aims to improve the code by avoiding circular dependencies and improving the import statements.

==
The above summary is generated by LLM with minor manual fixes. The following summary is by me.

The original import causes some issues when compiling DDP with compiled_autograd. The root cause of compilation failure is not identified but it is good to fix the lazy initialization, which indirectly fixes the compilation issues for DDP.
ghstack-source-id: 210051674
@exported-using-ghexport

Differential Revision: [D51857246](https://our.internmc.facebook.com/intern/diff/D51857246/)