Let all_reduce_coalesced accept one tensor as well
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115650

This diff introduces a change to the `all_reduce_coalesced` function in `distributed_c10d.py`. The function now accepts a single tensor as well as a list of tensors. This allows for more flexibility in the use of the function.


This is just a syntax sugar for the compiler to use `all_reduce_coalesced` without worrying  about converting the input to a list.
ghstack-source-id: 210051675
@exported-using-ghexport

Differential Revision: [D51433236](https://our.internmc.facebook.com/intern/diff/D51433236/)