[Inductor cutlass backend] Test and fix broadcasting of Aux / Bias edge cases
There are some edge cases concerning broadcasting of auxiliary inputs which need special treatment. The Cutlass
operators to broadcast auxiliary inputs require these inputs to be contiguous in the non-broadcasted dimension.

Most importantly, though, Pointwise nodes are able to implicitly reinterpret the memory layout of the Buffers they read from. In order to reliable way to fuse an additionally
loaded buffer as "Bias" argument, it is therefore neccessary to parse out of that
information a memory layout ( strides, offset and mapping to GEMM output dims ).

Added several tests to cover cases related to broadcasting of bias / aux inputs.

ghstack-source-id: d9b8abebcca14fac8e2d19dc95abb59b3e1a81ea
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115655