[C10D] Add waitForDumpOrTimeout to log on dump abandonment
Helps call attention to any cases where the dump actually times out.

The timeout is likely to hit if we run into slow stacktrace processing.

Log any exceptions encountered in the background thread, but don't raise
them- we're already willing to abandon the debug dump, and want to
proceed with our normal execution (in the case of dumppipe) or shutdown
process (when dumping happens on timeout and shutdown is already
initiated).

ghstack-source-id: 535a6e77b45495deac55e019ceacfb65c5b1f127
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115876