Preserve strides of custom Triton kernel args
Summary: Currently, we [`clone`](https://github.com/pytorch/pytorch/blob/19207b9183baced683d675c341b7da051cd16440/torch/_inductor/lowering.py#L5273) every `TensorBox` argument of custom Triton kernels while lowering them to the Inductor IR, during which the stride information of the kernel inputs is lost. This is problematic in the common case when the strides of a `torch.Tensor` argument are passed as scalars to a custom Triton kernel alongside the tensor itself (due to the underlying Triton code interpreting the tensors as raw pointers, so the contained stride semantics of the `torch.Tensor` is lost).

In this PR, we add an extended version of the existing [`clone` lowering](https://github.com/pytorch/pytorch/blob/19207b9183baced683d675c341b7da051cd16440/torch/_inductor/lowering.py#L2289)---`clone_preserve_reinterpret_view`---which carries over the `ir.ReinterpretVew` layers (if any) from the source `TensorBox` to the cloned one. The rationale behind adding a new function (and switching to it in the `triton_kernel_wrap` only for now) as opposed to extending the existing `clone` is keeping the semantics of the latter untouched, as it is a lowering of `torch.clone` (albeit incomplete, as the `memory_format` is currently ignored). Changing the existing `clone` would change the semantics which is not necessarily desirable in general. Open to suggestions, though.

Test Plan:

```
$ python test/dynamo/test_functions.py -k test_triton_kernel_strided_input
...
----------------------------------------------------------------------
Ran 1 test in 5.568s

OK
```

```
$ python test/dynamo/test_functions.py -k test_triton_kernel_strided_input_nonzero_offset
...
----------------------------------------------------------------------
Ran 1 test in 5.231s

OK
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: a74278b770a8a8af575f8a2c3527af6bb827a319
Pull Request resolved: https://github.com/pytorch/pytorch/pull/116219