[SparseCompressed] support csc layout for add sparse/dense.
`add` when passed one sparse and one dense argument  will error if the
sparse argument does not have  csr layout. This PR modifies the
underlying algorithm to be generic on the compressed dimension handling
both csr and csc. The functions are renamed to use the
`sparse_compressed` qualifier rather than `sparse_csr`

Fixes: #114807

ghstack-source-id: 62fda88d01e82092faa86597091d7d37ee5e7c4d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115433